fix: php example
