+++ php/src/generate.php
@@ -3,10 +3,7 @@ require_once "vendor/autoload.php";
 
 
 function generate() {
-    $libSearchPath = array(__DIR__, "/usr/local/lib", "/usr/lib", getenv("HOME")."/.local/lib");
-    $headerSearchPath = array(__DIR__, "/usr/local/include", "/usr/include", getenv("HOME")."/.local/include");
-
-    return (new FFIMe\FFIMe("libextism.".soext(), $headerSearchPath, $libSearchPath))
+    return (new FFIMe\FFIMe("libextism.".soext()))
         ->include("extism.h")
         ->showWarnings(false)
         ->codeGen('ExtismLib', __DIR__.'/ExtismLib.php');
